Added input variants without the formik connection to prevent "formik context is undefined" warning
