+++ package.json
@@ -53,15 +53,15 @@
     "@lessondesk/eslint-config": "1.4.1",
     "@lessondesk/material-icons": "^1.0.2",
     "@storybook/addon-a11y": "6.1.14",
-    "@storybook/addon-actions": "6.0.28",
-    "@storybook/addon-docs": "^6.0.21",
-    "@storybook/addon-knobs": "6.0.28",
-    "@storybook/addon-storyshots": "6.0.21",
-    "@storybook/addon-storyshots-puppeteer": "6.0.21",
-    "@storybook/addon-viewport": "6.0.28",
-    "@storybook/codemod": "^6.0.21",
+    "@storybook/addon-actions": "6.1.14",
+    "@storybook/addon-docs": "^6.1.14",
+    "@storybook/addon-knobs": "6.1.14",
+    "@storybook/addon-storyshots": "6.1.14",
+    "@storybook/addon-storyshots-puppeteer": "6.1.14",
+    "@storybook/addon-viewport": "6.1.14",
+    "@storybook/codemod": "^6.1.14",
     "@storybook/react": "6.1.14",
-    "@storybook/theming": "6.0.21",
+    "@storybook/theming": "6.1.14",
     "babel-eslint": "10.1.0",
     "babel-jest": "26.6.3",
     "babel-plugin-styled-components": "1.11.1",
@@ -90,7 +90,7 @@
     "react": "16.14.0",
     "react-dom": "16.14.0",
     "react-test-renderer": "16.13.1",
-    "storybook-readme": "5.0.8",
+    "storybook-readme": "5.0.9",
     "style-loader": "0.23.1",
     "styled-components": "5.2.1",
     "stylelint": "13.8.0",
