+++ src/additional-stories/radii/index.stories.mdx
@@ -9,7 +9,7 @@ import theme, { radiiAliases } from '../../config/theme'
 
 # Radii
 A list of the available radii properties. 
-Can be used in places where a component accepts a shadow,
+Can be used in places where a component accepts a radius,
 or can be accessed through the theme.
 
 <Flex flexDirection='column'>
